+++ app/src/main/AndroidManifest.xml
@@ -7,6 +7,7 @@
     <uses-permission android:name="android.permission.FOREGROUND_SERVICE_DATA_SYNC"/>
 
     <application
+        android:usesCleartextTraffic="false"
         android:name=".MyApplication"
         android:enableOnBackInvokedCallback="true"
         android:allowBackup="true"
