+++ app/src/main/kotlin/star/key13bb/emerald/MainActivity.kt
@@ -27,8 +27,8 @@ import star.key13bb.emerald.ui.screens.*
 import star.key13bb.emerald.ui.theme.EmeraldTheme
 
 class MainActivity : ComponentActivity() {
-    @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
     @OptIn(ExperimentalMaterial3WindowSizeClassApi::class)
+    @SuppressLint("UnusedMaterial3ScaffoldPaddingParameter")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
