+++ app/src/main/res/drawable/ic_launcher_anim_light.xml
@@ -36,7 +36,7 @@
                 <objectAnimator
                     android:duration="1000"
                     android:propertyName="pathData"
-                    android:valueFrom="M1,2c-4,0 -8,0.5 -8,4v9.5C4,17.43 5.57,19 7.5,19L6,20.5v0.5h2l2,-2h4l2,2h2v-0.5L16.5,19c1.93,0 3.5,-1.57 3.5,-3.5L20,6c0,-3.5 -4,-4 -8,-4zM2,4c3.51,0 4.96,0.48 5.57,1L9.43,5c0.61,-0.52 2.06,-1 5.57,-1zM0,7h5v3L6,10L6,7zM18,10.5c0,0.83 -0.67,1.5 -1.5,1.5h-9c-0.83,0 -1.5,-0.67 -1.5,-1.5L6,12h12v3.5zM25,10h-5L13,7h0v3z"
+                    android:valueFrom="M0,2c-4,0 -8,0.5 -8,4v9.5C4,17.43 5.57,19 7.5,19L6,20.5v0.5h2l2,-2h4l2,2h2v-0.5L16.5,19c1.93,0 3.5,-1.57 3.5,-3.5L20,6c0,-3.5 -4,-4 -8,-4zM2,4c3.51,0 4.96,0.48 5.57,1L9.43,5c0.61,-0.52 2.06,-1 5.57,-1zM0,7h5v3L6,10L6,7zM18,10.5c0,0.83 -0.67,1.5 -1.5,1.5h-9c-0.83,0 -1.5,-0.67 -1.5,-1.5L6,12h12v3.5zM25,10h-5L13,7h0v3z"
                     android:valueTo="M12,2c-4,0 -8,0.5 -8,4v9.5C4,17.43 5.57,19 7.5,19L6,20.5v0.5h2l2,-2h4l2,2h2v-0.5L16.5,19c1.93,0 3.5,-1.57 3.5,-3.5L20,6c0,-3.5 -4,-4 -8,-4zM12,4c3.51,0 4.96,0.48 5.57,1L6.43,5c0.61,-0.52 2.06,-1 5.57,-1zM6,7h5v3L6,10L6,7zM18,15.5c0,0.83 -0.67,1.5 -1.5,1.5h-9c-0.83,0 -1.5,-0.67 -1.5,-1.5L6,12h12v3.5zM18,10h-5L13,7h5v3z"
                     android:valueType="pathType"/>
             </set>
