+++ gradle/libs.versions.toml
@@ -5,11 +5,14 @@ datastore-preferences = "1.1.1"
 gtfs-validator-model = "5.0.0"
 kotlinx-coroutines-android = "1.9.0"
 lifecycle = "2.9.0-alpha05"
+material3-window-size = "1.4.0-alpha02"
+sqlite-jdbc-version = "3.46.1.3"
 
 [libraries]
 
 androidx-compose-bom = { group = "androidx.compose", name = "compose-bom", version = "2024.10.00" }
-firebase-bom = { module = "com.google.firebase:firebase-bom", version = "33.4.0" }
+androidx-material3-window-size = { module = "androidx.compose.material3:material3-window-size-class", version.ref = "material3-window-size" }
+firebase-bom = { module = "com.google.firebase:firebase-bom", version = "33.5.0" }
 
 androidx-activity-compose = { group = "androidx.activity", name = "activity-compose", version = "1.10.0-alpha02" }
 androidx-core-ktx = { group = "androidx.core", name = "core-ktx", version = "1.15.0-rc01" }
@@ -33,7 +36,7 @@ androidx-ui-tooling-preview = { group = "androidx.compose.ui", name = "ui-toolin
 datastore-preferences-core = { module = "androidx.datastore:datastore-preferences-core", version.ref = "datastore-preferences-core-version" }
 
 firebase-analytics = { module = "com.google.firebase:firebase-analytics", version = "22.1.2" }
-firebase-crashlytics = { module = "com.google.firebase:firebase-crashlytics", version = "19.2.0" }
+firebase-crashlytics = { module = "com.google.firebase:firebase-crashlytics", version = "19.2.1" }
 
 gtfs-realtime-bindings = { module = "org.mobilitydata:gtfs-realtime-bindings", version = "0.0.8" }
 gtfs-validator-model = { module = "org.mobilitydata.gtfs-validator:gtfs-validator-model", version.ref = "gtfs-validator-model" }
@@ -44,6 +47,7 @@ protobuf-kotlin = { module = "com.google.protobuf:protobuf-java", version = "4.2
 
 androidx-ui-test-manifest = { group = "androidx.compose.ui", name = "ui-test-manifest", version.ref = "androidx-ui" }
 androidx-ui-tooling = { group = "androidx.compose.ui", name = "ui-tooling", version.ref = "androidx-ui" }
+xerial-sqlite-jdbc = { module = "org.xerial:sqlite-jdbc", version.ref = "sqlite-jdbc-version" }
 
 
 [plugins]
